enh(arch): move metrics related code into more suitable folders
