+++ centreon/src/Core/Metric/Domain/Model/Metric.php
@@ -21,20 +21,31 @@
 
 declare(strict_types=1);
 
-namespace Core\Infrastructure\RealTime\Api\DownloadPerformanceMetrics;
+namespace Core\Metric\Domain\Model;
 
-use Core\Application\Common\UseCase\AbstractPresenter;
-use Core\Application\RealTime\UseCase\FindPerformanceMetrics\FindPerformanceMetricResponse;
-
-class DownloadPerformanceMetricsPresenter extends AbstractPresenter
+class Metric
 {
     /**
-     * {@inheritDoc}
-     *
-     * @param FindPerformanceMetricResponse $data
+     * @param int $id
+     * @param string $name
      */
-    public function present(mixed $data): void
+    public function __construct(private int $id, private string $name)
     {
-        parent::present($data);
+    }
+
+    /**
+     * @return int
+     */
+    public function getId(): int
+    {
+        return $this->id;
+    }
+
+    /**
+     * @return string
+     */
+    public function getName(): string
+    {
+        return $this->name;
     }
 }
